pyproject.toml
[project]
name = "iso26262-docgen"
version = "0.1.0"
description = "Generate ISO 26262 ↔ Rust mapping DOCX from Markdown + YAML tables"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "jsonschema>=4.21.0",
  "sphinx==8.2.3",
  "myst-parser==4.0.1",
]

[dependency-groups]
dev = []
